refactor: export apis with out dependencies of react (#187)
* refactor: export apis with out dependencies of react

* fix: error message when unmount
