+++ packages/icestark-module/src/index.ts
@@ -1,11 +1,11 @@
 import {
-  MicroModule,
   registerModules,
   getModules,
   mountModule,
   unmoutModule,
   clearModules,
 } from './modules';
+import MicroModule from './MicroModule';
 
 export {
   MicroModule,
