+++ packages/icestark-module/src/modules.tsx
@@ -1,9 +1,7 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
 import Sandbox, { SandboxProps, SandboxContructor } from '@ice/sandbox';
 import ModuleLoader, { StarkModule } from './loader';
 
-type ISandbox = boolean | SandboxProps | SandboxContructor;
+export type ISandbox = boolean | SandboxProps | SandboxContructor;
 
 let globalModules = [];
 let importModules = {};
@@ -21,33 +19,18 @@ export const clearModules = () => {
   importModules = {};
   moduleLoader.clearTask();
 };
-
-/**
- * Render Component, compatible with Component and <Component>
- */
-export function renderComponent(Component: any, props = {}): React.ReactElement {
-  return React.isValidElement(Component) ? (
-    React.cloneElement(Component, props)
-  ) : (
-    // eslint-disable-next-line react/jsx-filename-extension
-    <Component {...props} />
-  );
-}
-
 /**
  * support react module render
  */
-const defaultMount = (Component: any, targetNode: HTMLElement, props?: any) => {
-  console.warn('Please set mount, try run react mount function');
-  ReactDOM.render(renderComponent(Component, props), targetNode);
+const defaultMount = () => {
+  console.error('[icestark module] Please export mount function');
 };
 
 /**
  * default unmount function
  */
-const defaultUnmount = (targetNode: HTMLElement) => {
-  console.warn('Please set unmount, try run react unmount function');
-  ReactDOM.unmountComponentAtNode(targetNode);
+const defaultUnmount = () => {
+  console.error('[icestark module] Please export unmount function');
 };
 
 function createSandbox(sandbox: ISandbox) {
@@ -194,91 +177,3 @@ export const unmoutModule = (targetModule: StarkModule, targetNode: HTMLElement)
   return unmount(targetNode);
 };
 
-/**
- * default render compoent, mount all modules
- */
-export class MicroModule extends React.Component<any, { loading: boolean }> {
-  private moduleInfo = null;
-
-  private mountNode = null;
-
-  private unmout = false;
-
-  static defaultProps = {
-    loadingComponent: null,
-    handleError: () => {},
-  };
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      loading: false,
-    };
-  }
-
-  componentDidMount() {
-    this.mountModule();
-  }
-
-  componentDidUpdate(prevProps) {
-    if (prevProps.moduleInfo !== this.props.moduleInfo || prevProps.moduleName !== this.props.moduleName) {
-      this.mountModule();
-    }
-  }
-
-  componentWillUnmount() {
-    unmoutModule(this.moduleInfo, this.mountNode);
-    this.unmout = true;
-  }
-
-  async mountModule() {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { sandbox, moduleInfo, wrapperClassName, wrapperStyle, loadingComponent, handleError, ...rest } = this.props;
-    this.moduleInfo = moduleInfo || getModules().filter(m => m.name === this.props.moduleName)[0];
-    if (!this.moduleInfo) {
-      console.error(`Can't find ${this.props.moduleName} module in modules config`);
-      return;
-    }
-    this.setState({ loading: true });
-    try {
-      const { mount, component } =  await loadModule(this.moduleInfo, sandbox);
-      this.setState({ loading: false });
-      if (mount && component) {
-        if (this.unmout) {
-          unmoutModule(this.moduleInfo, this.mountNode);
-        } else {
-          mount(component, this.mountNode, rest);
-        }
-      }
-    } catch (err) {
-      this.setState({ loading: false });
-      handleError(err);
-    }
-  }
-
-  render() {
-    const { loading } = this.state;
-    const { wrapperClassName, wrapperStyle, loadingComponent } = this.props;
-    return loading ? loadingComponent
-      : <div className={wrapperClassName} style={wrapperStyle} ref={ref => this.mountNode = ref} />;
-  }
-};
-
-/**
- * Render Modules, compatible with Render and <Render>
- */
-export default function renderModules(modules: StarkModule[], render: any, componentProps?: any, sandbox?: ISandbox): React.ReactElement {
-  // save match app modules in global
-  registerModules(modules);
-
-  if (render) {
-    return renderComponent(render, {
-      modules,
-      ...componentProps,
-      sandbox,
-    });
-  }
-
-  console.warn('Please set render Component, try use MicroModule and mount first module');
-  return <MicroModule moduleName={modules[0]?.name} {...componentProps} />;
-};
